Feat : BookService 단위 테스트 (#355)
BookRepository 단위 테스트

+++ backend/src/main/java/com/dongnebook/global/security/auth/filter/JwtAuthenticationFilter.java
@@ -51,7 +51,7 @@ public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilte
         String refreshToken = tokenDto.getRefreshToken();
 
         ResponseCookie cookie = ResponseCookie.from("refreshToken", refreshToken)
-            .maxAge(7 * 24 * 60 * 60)
+            .maxAge((long)7 * 24 * 60 * 60)
             .path("/")
             .secure(true)
             .sameSite("None")
